Move `NumericField` component to a `PageComponents` file.
