+++ client/src/main/kotlin/io/spine/examples/shareaware/client/PageComponents.kt
@@ -42,6 +42,8 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -52,6 +54,7 @@ import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.dp
 import io.spine.examples.shareaware.MoneyCalculator
 import io.spine.examples.shareaware.client.MoneyExtensions.asReadableString
+import io.spine.examples.shareaware.client.StringExtensions.validateNumber
 import io.spine.examples.shareaware.client.payment.Popup
 import io.spine.examples.shareaware.client.payment.PopupConfig
 import io.spine.money.Money
@@ -161,6 +164,32 @@ public fun SearchField(
             )
         )
     }
+}
+
+/**
+ * Displays the input component accepting only the numeric values.
+ *
+ * @param placeholder the label to be displayed inside the input container
+ * @param onChange the callback that is triggered when the input's value changes
+ */
+@Composable
+public fun NumericInput(
+    placeholder: String,
+    onChange: (String) -> Unit
+) {
+    val input = remember { mutableStateOf("") }
+    val onChange: (String) -> Unit = {
+        if (it.validateNumber()) {
+            input.value = it
+            onChange(it)
+        }
+    }
+    Input(
+        value = input.value,
+        onChange = onChange,
+        placeholder = placeholder,
+        isError = false
+    )
 }
 
 /**
